% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\alias{print.describe}
\title{Descriptive statistics}
\usage{
describe(data, charts = TRUE, ...)

\method{print}{describe}(x, ...)
}
\arguments{
\item{data}{a matrix or a data.frame.}

\item{\dots}{additional arguments to be passed to the low level functions.}
}
\description{
Descriptive statistics of a data frame.
}
\examples{

describe(iris)
describe(dplyr::group_by(iris, Species))

}
