% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_glmnet.R
\name{ic.glmnet}
\alias{ic.glmnet}
\title{Information criteria for glmnet models}
\usage{
ic.glmnet(object, ...)
}
\arguments{
\item{object}{an object of class \code{'glmnet'} obtained from \code{\link[glmnet]{glmnet}} function.}

\item{\dots}{additional arguments to be passed to the low level functions.}
}
\value{
A list containing the following elements:
\itemize{
  \item family
  \item alpha
  \item lambda
  \item dev
  \item dev.ratio
  \item loglik
  \item df
  \item aic
  \item aicc
  \item bic
}
}
\description{
Calculates information criteria (AIC, AICc, BIC) for Ridge, Lasso, and
ElasticNet regression models.
}
\examples{
X = matrix(rnorm(100 * 20), 100, 20)
b = c(runif(5, min = -1, max = 1), rep(0,15))
y = X \%*\% b + rnorm(100)
mod_ridge = glmnet::glmnet(X, y, alpha = 0, family = "gaussian")
ic = ic.glmnet(mod_ridge)
str(ic)
with(ic, { plot(log(lambda), aic); abline(v = lambda[which.min(aic)], lty = 2) })
with(ic, { plot(log(lambda), bic); abline(v = lambda[which.min(bic)], lty = 2) })

mod_lasso = glmnet::glmnet(X, y, alpha = 1, family = "gaussian")
plot(mod_lasso, xvar = "lambda")
ic = ic.glmnet(mod_lasso)
str(ic)
with(ic, { plot(log(lambda), aic); abline(v = log(lambda[which.min(aic)]), lty = 2) })
with(ic, { plot(log(lambda), bic); abline(v = log(lambda[which.min(bic)]), lty = 2) })

}
