% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclustda.R
\docType{data}
\name{mclustda}
\alias{mclustda}
\title{Code to support caret interface for Gaussian mixtures in
classification from mclust package}
\format{
An object of class \code{list} of length 11.
}
\usage{
mclustda
}
\description{
A list of elements used by \code{\link[caret]{train}}
for training and hyper-parameter tuning of Gaussian mixture models 
in classification fitted by \code{\link[mclust]{MclustDA}}.
}
\examples{
str(mclustda, give.attr = FALSE)
\dontrun{
mod1 = train(x = iris[,1:4], y = iris$Species, 
            method = mclustda, 
            trControl = trainControl(method = "cv", number = 10))
mod1
plot(mod1)

mod2 = train(x = iris[,1:4], y = iris$Species, 
             method = mclustda, 
             metric = "BrierScore", maximize = FALSE,
             trControl = trainControl(method = "cv", number = 10,
                                      summaryFunction = mclassSummary,
                                      classProbs = TRUE))
mod2
plot(mod2)

mod2$results$G = as.factor(mod2$results$G)
ggplot(mod2$results, aes(y = .data[[mod2$metric]],
                        x = modelName, 
                        col = G, 
                        shape = modelType)) +
  geom_point()
}
}
\keyword{datasets}
