% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_addson.R
\name{prcompRecon}
\alias{prcompRecon}
\title{Principal components data reconstruction}
\usage{
prcompRecon(object, npcs = 2, pcs = NULL, ...)
}
\arguments{
\item{object}{an object returned by `prcomp()` function.}

\item{npcs}{the number of principal components to use in data reconstruction}

\item{pcs}{specify which principal components to use in data reconstruction. If provided it takes precedence over `npcs`}

\item{\dots}{additional arguments to be passed to the low level functions.}
}
\value{
The reconstructed data matrix.
}
\description{
Return the data matrix reconstructed using a subset of 
principal components.
}
\examples{

pairs(iris[,1:4], gap = 0)
PCA = prcomp(iris[,1:4])
X = prcompRecon(PCA, 2)
pairs(X, gap = 0)
X = prcompRecon(PCA, pcs = c(1,4))
pairs(X, gap = 0)

}
