% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssim.R
\name{ssim}
\alias{ssim}
\title{Structural similarity index}
\usage{
ssim(x, y, L = 255)
}
\arguments{
\item{x}{= a matrix of values for a segmented/approximated image}

\item{y}{= a matrix of values for the original image}

\item{L}{= dynamic range of the pixel-values, for a 8-bit image 
$L = 2^8-1 = 255$.}
}
\value{
The computed SSIM value.
}
\description{
Computes the Structural Similarity Index Measure (SSIM) is a
widely adopted index for measuring the similarity between two images by
evaluating the perceived change in structural information.
SSIM \eqn{\in}{in} [0,1], with SSIM = 1 indicating a perfect match with
the original image. For values larger than 0.95 two images are
considered indistinguishable to the human eye.
}
